app/requirements.txt
asgiref==3.2.7
Babel==2.8.0
blinker==1.4
click==7.1.1
Django==3.0.5
django-qrcode==0.3
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.1.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
qrcode==6.1
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
Werkzeug==1.0.0
